+++ changelog.md
@@ -9,3 +9,5 @@
 - [index] minor: use `int-pack-vec` glsl and js to power the default picking behavior. picking will yield the feature id, which is produced by `prepare._splitSort`, which happens on prepare.update`. there is also a default `pick` function that wraps the `map.pick` as a convinience, since we are picking 2px wide, and then we unpack and give the caller a feature index and feature type.
 
 - [prepare] minor: replace Prepare.prototype.getOpacity with a function exported from `georender-style2png` which pulls opacity. this centralizes all style texture writing and reading into one place. same with finding zindex values.
+
+- [prepare] minor: add an `extend : (attributes, props) => void` to the `prepare` options, this allows for the ability to define additional attribute data that should be spread across T and P keys upon zoom level update (`prepare.update(zoom)`).